Rename register feed usecase directory
